requirements.txt
deep_translator
Pillow